Code has been refactored, tests have been added. Needs javaDocs.

+++ src/main/java/pl/mg/currConv/model/UserInputDto.java
@@ -1,16 +1,18 @@
 package pl.mg.currConv.model;
 
+import java.math.BigDecimal;
+
 public class UserInputDto {
-    private final double amount;
+    private final BigDecimal amount;
 
     private final String currencyCode;
 
-    public UserInputDto(final double amount, final String currencyCode) {
+    public UserInputDto(final BigDecimal amount, final String currencyCode) {
         this.amount = amount;
         this.currencyCode = currencyCode;
     }
 
-    public double getAmount() { return amount; }
+    public BigDecimal getAmount() { return amount; }
 
     public String getCurrencyCode() { return currencyCode; }
 }
